lab 5 task (пятая лабораторная с решением задачи)
Требовалось создать Policy для Blog (заменить middleware в роутах на политику)

+++ app/Http/Controllers/BlogController.php
@@ -16,10 +16,12 @@ class BlogController extends Controller
         return view('blog/index', compact('posts'));
     }
     public function create() {
+        $this->authorize('create', BlogPost::class);
         Spy::process(request());
         return view('blog/create');
     }
     public function store(BlogRequest $request) {
+        $this->authorize('create', BlogPost::class);
         $newPost = $request->validated();
         Spy::process($request);
         BlogService::create($newPost);
@@ -31,26 +33,31 @@ class BlogController extends Controller
         return view('blog/show', compact('model'));
     }
     public function edit($id) {
-        Spy::process(request());
         $model = BlogPost::find($id);
+        $this->authorize('update', $model);
+        Spy::process(request());
         return view('blog/edit', compact('model'));
     }
     public function update(BlogRequest $request, $id) {
+        $this->authorize('update', BlogPost::find($id));
         $updateData = $request->validated();
         Spy::process($request);
         BlogService::update($updateData, $id);
         return redirect() -> route('blog.show', $id);
     }
     public function destroy($id) {
+        $this->authorize('destroy', BlogPost::find($id));
         Spy::process(request());
         BlogService::delete($id);
         return redirect() -> route('blog.index');
     }
     public function file_upload() {
+        $this->authorize('file_upload', BlogPost::class);
         Spy::process(request());
         return view('blog/file_upload');
     }
     public function file_upload_update() {
+        $this->authorize('file_upload', BlogPost::class);
         $csv = request()->validate(['csv' => 'required'])['csv'];
         Spy::process(request());
         (new BlogCsvUploaderService($csv))->file_upload();
